Add files via upload
The above file contains the project related files including R code, report of the project in the form of PDF and the dataset used for the analysis. The project is about analysis of the bank customers that when given a bank loan will they default or not. The project uses Decision Tree and Random Forest algorithms to make predictions.